Create a pizza recipe page
The landing page of the recipies site should be filled with different recipes.
The first of which is pizza.html. The landing page was updated to create a link
to pizza.html. pizza.html was filled with HTML boilerplate and a title. A link to
return to the homepage was also created.