Create a recipe for waffles
This commit add a wallfes recipe on its own page with the same code and differnt copy.
The return to home button on pizza.html was also moved to the bottom of the page for visual purposes.
Stock images for waffles and tacos to be used later were both added.